Simplified transform result logic

+++ src/main/java/org/documents/documents/helper/impl/RunTransformHelperImpl.java
@@ -2,18 +2,20 @@ package org.documents.documents.helper.impl;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.documents.documents.db.entity.RenditionEntity;
 import org.documents.documents.db.repository.ContentRepository;
-import org.documents.documents.file.FileReference;
-import org.documents.documents.file.FileStoreRegistry;
-import org.documents.documents.file.TransformFileStore;
+import org.documents.documents.file.*;
 import org.documents.documents.helper.*;
+import org.documents.documents.model.ContentAndRenditionEntities;
 import org.documents.documents.model.transform.TransformResult;
 import org.documents.documents.transform.Transform;
 import org.documents.documents.transform.TransformRegistry;
 import org.springframework.stereotype.Component;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @Component
@@ -27,32 +29,56 @@ public class RunTransformHelperImpl implements RunTransformHelper {
     private final ContentRepository contentRepository;
 
     @Override
-    public void runTransform(UUID contentUuid, FileReference fileReference, String sourceMimeType, String targetMimeType) {
-        try {
-            final Optional<Transform> foundTransform = transformRegistry.getTransform(sourceMimeType, targetMimeType);
-            if (foundTransform.isPresent()) {
-                final Transform transform = foundTransform.get();
-                final UUID uuid = fileStoreRegistry.copyToTransformFileStore(fileReference, sourceMimeType);
-                final TransformResult result = transform.transform(uuid, sourceMimeType, targetMimeType);
-                if (result.success()) {
-                    onSuccess(contentUuid, uuid, targetMimeType);
+    public void runTransform(UUID contentUuid, String targetMimeType) {
+        final ContentAndRenditionEntities entities = Objects.requireNonNull(renditionHelper.getContentAndRenditions(contentUuid).block());
+        final String sourceMimeType = entities.getContentEntity().getMimeType();
+        if(!entities.getRenditionEntitiesByMimeType().containsKey(targetMimeType)) {
+            try {
+                final Optional<Transform> foundTransform = transformRegistry.getTransform(sourceMimeType, targetMimeType);
+                if (foundTransform.isPresent()) {
+                    final Transform transform = foundTransform.get();
+                    final TypedFileReference bestFileReference = getBestFileReference(entities, transform, targetMimeType);
+                    final UUID uuid = fileStoreRegistry.copyToTransformFileStore(bestFileReference.toFileReference(), bestFileReference.mimeType());
+                    final TransformResult result = transform.transform(uuid, bestFileReference.mimeType(), targetMimeType);
+                    if (result.isSuccess()) {
+                        onSuccess(contentUuid, uuid, result.getOutputMimeTypes());
+                    } else {
+                        onFailure(contentUuid, result.getErrorMessage());
+                    }
+                    transformFileStore.delete(uuid, sourceMimeType);
                 } else {
-                    onFailure(contentUuid, result.message());
+                    log.warn("no transform found for {} -> {}", sourceMimeType, targetMimeType);
                 }
-                transformFileStore.delete(uuid, sourceMimeType);
-            } else {
-                log.warn("no transform found for {} -> {}", sourceMimeType, targetMimeType);
+            } catch(Exception e) {
+                log.error("failed to run transform", e);
             }
-        } catch(Exception e) {
-            log.error("failed to run transform", e);
         }
     }
+    private TypedFileReference getBestFileReference(ContentAndRenditionEntities entities, Transform transform, String targetMimeType) {
+        final List<String> preferredMimeTypes = transform.getPreferredSourceMimeTypes(targetMimeType);
+        for(String preferredMimeType : preferredMimeTypes) {
+            if (entities.getContentEntity().getMimeType().equals(preferredMimeType)) {
+                return new TypedFileReference(FileStoreType.CONTENT, UUID.fromString(entities.getContentEntity().getUuid()), entities.getContentEntity().getMimeType());
+            } else {
+                final RenditionEntity renditionEntity = entities.getRenditionEntitiesByMimeType().get(preferredMimeType);
+                if(renditionEntity != null) {
+                    return new TypedFileReference(FileStoreType.RENDITION, UUID.fromString(renditionEntity.getUuid()), renditionEntity.getMimeType());
+                }
+            }
+        }
+        return new TypedFileReference(FileStoreType.CONTENT, UUID.fromString(entities.getContentEntity().getUuid()), entities.getContentEntity().getMimeType());
+    }
 
-    private void onSuccess(UUID contentUuid, UUID transformedUuid, String targetMimeType) {
-        contentRepository.findByUuid(contentUuid.toString())
-                .flatMap(contentEntity -> renditionHelper.storeRendition(contentEntity, targetMimeType, transformedUuid))
-                .block();
-        log.debug("Rendition successfully stored {} : {}", contentUuid, targetMimeType);
+    private void onSuccess(UUID contentUuid, UUID transformedUuid, Set<String> outputMimeTypes) {
+        final List<RenditionEntity> renditions = Objects.requireNonNull(contentRepository.findByUuid(contentUuid.toString())
+                .flatMapMany(contentEntity ->
+                        Flux.fromIterable(outputMimeTypes).flatMap(outputMimeType -> renditionHelper.storeRendition(contentEntity, outputMimeType, transformedUuid))
+                ).collect(Collectors.toList())
+                .switchIfEmpty(Mono.just(Collections.emptyList()))
+                .block());
+        for(RenditionEntity renditionEntity : renditions) {
+            log.debug("Rendition {} for mime type {} of content UUID {} successfully stored", renditionEntity.getUuid(), renditionEntity.getMimeType(), contentUuid);
+        }
     }
 
     private void onFailure(UUID uuid, String message) {
